fix(app): 빌드 통과 — 타입/opt-in 수정 + gradle wrapper + compileSdk 34
- LinearProgressIndicator progress: Double→Float (toFloat()) 3곳
- StockApp: @OptIn(ExperimentalFoundationApi) — HorizontalPager opt-in
- compileSdk/targetSdk 35→34 (설치된 SDK)
- gradle wrapper(gradlew/jar) 추가 → CI·타 환경 빌드 가능
- 결과: BUILD SUCCESSFUL, app-debug.apk 8.6MB 생성

+++ android/app/src/main/java/com/oldbutrich/stockpick/ui/StockDetailScreen.kt
@@ -177,7 +177,7 @@ fun SummaryCard(stock: Recommendation) {
 
             // 종합 점수 바
             LinearProgressIndicator(
-                progress = { (stock.overallScore / 100f).coerceIn(0f, 1f) },
+                progress = { (stock.overallScore.toFloat() / 100f).coerceIn(0f, 1f) },
                 modifier = Modifier
                     .fillMaxWidth()
                     .height(8.dp)
@@ -288,7 +288,7 @@ fun CriterionCard(criterion: CriterionResult, isPrimary: Boolean) {
             // 점수 바
             Spacer(Modifier.height(10.dp))
             LinearProgressIndicator(
-                progress = { (criterion.score / 100f).coerceIn(0f, 1f) },
+                progress = { (criterion.score.toFloat() / 100f).coerceIn(0f, 1f) },
                 modifier = Modifier
                     .fillMaxWidth()
                     .height(5.dp)
